Track properties and components at intance level
Instead of keeping class levels of a component, just use the DSL class
methods to generate instance methods that track the sub-components and
properties in an instance-level store.

This allows us to provide an instance-level API to extend a component
with custom properties/sub-components that aren't defined in the
library.