No more setting the method name on definition time
This is something that concerns objects *mounting* properties /
parameters, not the property / parameter itself, since different objects
could potentially have different method names for the same property.

For example, a property that is defined at most once in a certain
component, but multiple times in another component, would have a
singular name for one, and a plural name for the other.