Add timezone detection for time-based properties
When a property that includes Property::TimezoneDetection receives a
value, it will try to detect in which timezone it is, and set the TZID
parameter if appropriate for that value.

If the Time instance already carries timezone information (for example,
an instance of ActiveSupport::TimeWithZone) then we will use that.

If it doesn't, we will use TZInfo's database to find a Timezone that
best matches the information in the time object.

UTC times are ignored, as the RFC specifies that UTC timezones must not
have any timezone information past the Z identifier.